Log entry for end of day:
When I began this, I was not using Pandas, but today I was able to update my code to use Pandas. I found it much easier than I expected, and got most of my functions updated accordingly. I was able to get the user input to select which row to return so that was a big achievement.