Log entry for end of day 09/30/2023
I’m adding a log entry for yesterday as I didn’t do that. As of yesterday, I was able to create a function that calculates the puppy’s age. I was also able to create my separate functions that read my csv files and write to them. Lastly, I was able to print an image from my folder.